--- .env.example
@@ -1,3 +1,0 @@
-REPODIR=/path/to/wf-psf/repo
-DATADIR=${REPODIR}/data/coherent_euclid_dataset
-WORKDIR=/path/to/output/dir
