+++ wf_psf/metrics/metrics.py
@@ -2,9 +2,9 @@ import numpy as np
 import tensorflow as tf
 import matplotlib.pyplot as plt
 import galsim as gs
-import wf_psf.utils as utils
-from wf_psf.tf_psf_field import build_PSF_model
-from wf_psf import tf_psf_field as psf_field
+import wf_psf.utils.utils as utils
+from tf_psf_field import build_PSF_model
+from . import tf_psf_field as psf_field
 from wf_psf import SimPSFToolkit as SimPSFToolkit
