+++ wf_psf/sandbox.py
@@ -1,8 +1,8 @@
 import numpy as np
 import tensorflow as tf
 import matplotlib.pyplot as plt
-from wf_psf.utils import generate_packed_elems
-from wf_psf.tf_psf_field import build_PSF_model
+from wf_psf.utils.utils import generate_packed_elems
+from tf_psf_field import build_PSF_model
 
 """ Place to temporarily save some useful functions.
 """
