+++ wf_psf/training/tf_alt_psf_models.py
@@ -1,7 +1,7 @@
 import numpy as np
 import tensorflow as tf
-from wf_psf.tf_layers import TF_poly_Z_field, TF_zernike_OPD, TF_batch_poly_PSF
-from wf_psf.tf_layers import TF_NP_poly_OPD, TF_batch_mono_PSF
+from tf_layers import TF_poly_Z_field, TF_zernike_OPD, TF_batch_poly_PSF
+from tf_layers import TF_NP_poly_OPD, TF_batch_mono_PSF
 
 
 class TF_SemiParam_field_l2_OPD(tf.keras.Model):
